Move formatAddrShort, update imports, and first pass at copy changes
