+++ client/scripts/controllers/chain/substrate/treasury.ts
@@ -5,7 +5,6 @@ import { BalanceOf, Permill, BlockNumber } from '@polkadot/types/interfaces';
 import { DeriveBalancesAccount } from '@polkadot/api-derive/types';
 import { stringToU8a, u8aToHex } from '@polkadot/util';
 import { IApp } from 'state';
-import { formatAddressShort } from 'helpers';
 import {
   ISubstrateTreasuryProposal,
   SubstrateCoin
@@ -14,6 +13,7 @@ import { ProposalModule } from 'models';
 import { SubstrateTypes } from '@commonwealth/chain-events';
 import SubstrateChain from './shared';
 import SubstrateAccounts, { SubstrateAccount } from './account';
+import { formatAddressShort } from '../../../../../shared/utils';
 import { SubstrateTreasuryProposal } from './treasury_proposal';
 
 class SubstrateTreasury extends ProposalModule<
