+++ client/scripts/helpers/index.ts
@@ -206,15 +206,6 @@ export function formatDuration(duration: moment.Duration, includeSeconds = true)
   ].join('');
 }
 
-export function formatAddressShort(address: string, chain: string) {
-  if (!address) return;
-  if (chain === 'near') {
-    return `@${address}`;
-  } else {
-    return `${address.slice(0, 5)}…`;
-  }
-}
-
 export function formatProposalHashShort(hash: string) {
   if (!hash) return;
   return `${hash.slice(0, 8)}…`;
